deleted theme nofier
