+++ packages/mq_app_theme/lib/mq_app_theme.dart
@@ -1,4 +1,4 @@
-/// My Quran app theme
-library mq_app_theme;
+// /// My Quran app theme
+// library mq_app_theme;
 
-export 'src/mq_app_theme.dart';
+// export 'src/mq_app_theme.dart';
