+++ packages/mq_app_theme/lib/src/app_repository/app_repository.dart
@@ -1,11 +1,11 @@
-import 'package:mq_app_theme/mq_app_theme.dart';
+// import 'package:mq_app_theme/mq_app_theme.dart';
 
-abstract class AppRepository {
-  String appVersion();
+// abstract class AppRepository {
+//   String appVersion();
 
-  CustomTheme getInitialThemeState();
+//   CustomTheme getInitialThemeState();
 
-  Future<void> saveThemePrimaryColor(int index);
+//   Future<void> saveThemePrimaryColor(int index);
 
-  Future<void> saveThemeMode({required bool isDark});
-}
+//   Future<void> saveThemeMode({required bool isDark});
+// }
