+++ packages/mq_app_theme/lib/src/app_repository/app_repositoty_impl.dart
@@ -1,29 +1,29 @@
-import 'package:flutter/material.dart';
-import 'package:mq_app_theme/mq_app_theme.dart';
+// import 'package:flutter/material.dart';
+// import 'package:mq_app_theme/mq_app_theme.dart';
 
-@immutable
-final class AppRepositoryImpl implements AppRepository {
-  const AppRepositoryImpl(
-    this.dataSource,
-  );
+// @immutable
+// final class AppRepositoryImpl implements AppRepository {
+//   const AppRepositoryImpl(
+//     this.dataSource,
+//   );
 
-  final AppLocalDataSource dataSource;
+//   final AppLocalDataSource dataSource;
 
-  @override
-  String appVersion() => dataSource.appVersion();
+//   @override
+//   String appVersion() => dataSource.appVersion();
 
-  @override
-  CustomTheme getInitialThemeState() {
-    return dataSource.initialTheme();
-  }
+//   @override
+//   CustomTheme getInitialThemeState() {
+//     return dataSource.initialTheme();
+//   }
 
-  @override
-  Future<void> saveThemeMode({required bool isDark}) {
-    return dataSource.saveThemeMode(isDark: isDark);
-  }
+//   @override
+//   Future<void> saveThemeMode({required bool isDark}) {
+//     return dataSource.saveThemeMode(isDark: isDark);
+//   }
 
-  @override
-  Future<void> saveThemePrimaryColor(int index) {
-    return dataSource.saveThemePrimaryColor(index);
-  }
-}
+//   @override
+//   Future<void> saveThemePrimaryColor(int index) {
+//     return dataSource.saveThemePrimaryColor(index);
+//   }
+// }
