+++ packages/mq_app_theme/lib/src/app_repository/read_theme_repository_impl.dart
@@ -1,19 +1,19 @@
-import 'package:flutter/material.dart';
-import 'package:mq_app_theme/mq_app_theme.dart';
+// import 'package:flutter/material.dart';
+// import 'package:mq_app_theme/mq_app_theme.dart';
 
-@immutable
-final class ReadThemeRepositoryImpl implements ReadThemeRepository {
-  const ReadThemeRepositoryImpl(this.localThemeDataSource);
+// @immutable
+// final class ReadThemeRepositoryImpl implements ReadThemeRepository {
+//   const ReadThemeRepositoryImpl(this.localThemeDataSource);
 
-  final LocalThemeDataSourceImpl localThemeDataSource;
+//   final LocalThemeDataSourceImpl localThemeDataSource;
 
-  @override
-  ReadThemeData getInitialThemeState() {
-    return localThemeDataSource.getInitialThemeState();
-  }
+//   @override
+//   ReadThemeData getInitialThemeState() {
+//     return localThemeDataSource.getInitialThemeState();
+//   }
 
-  @override
-  Future<void> saveThemeState(ReadThemeData themeState) {
-    return localThemeDataSource.saveThemeState(themeState);
-  }
-}
+//   @override
+//   Future<void> saveThemeState(ReadThemeData themeState) {
+//     return localThemeDataSource.saveThemeState(themeState);
+//   }
+// }
