+++ packages/mq_app_theme/lib/src/app_theme/colors/app_colors.dart
@@ -1,13 +1,13 @@
-import 'package:flutter/material.dart';
+// import 'package:flutter/material.dart';
 
-@immutable
-final class AppColors {
-  const AppColors._();
+// @immutable
+// final class AppColors {
+//   const AppColors._();
 
-  static const white = Colors.white;
-  static const black = Colors.black;
-  static const red = Colors.red;
-  static const yellow = Colors.yellow;
-  static const green = Colors.green;
-  static const greyBlack = Color.fromARGB(255, 28, 28, 28);
-}
+//   static const white = Colors.white;
+//   static const black = Colors.black;
+//   static const red = Colors.red;
+//   static const yellow = Colors.yellow;
+//   static const green = Colors.green;
+//   static const greyBlack = Color.fromARGB(255, 28, 28, 28);
+// }
