+++ packages/mq_app_theme/lib/src/source/app_local_data_source.dart
@@ -1,11 +1,11 @@
-import 'package:mq_app_theme/mq_app_theme.dart';
+// import 'package:mq_app_theme/mq_app_theme.dart';
 
-abstract class AppLocalDataSource {
-  String appVersion();
+// abstract class AppLocalDataSource {
+//   String appVersion();
 
-  CustomTheme initialTheme();
+//   CustomTheme initialTheme();
 
-  Future<void> saveThemePrimaryColor(int index);
+//   Future<void> saveThemePrimaryColor(int index);
 
-  Future<void> saveThemeMode({required bool isDark});
-}
+//   Future<void> saveThemeMode({required bool isDark});
+// }
