+++ packages/mq_app_theme/lib/src/source/mock/app_local_data_source_mock.dart
@@ -1,19 +1,19 @@
-import 'package:flutter/material.dart';
-import 'package:mq_app_theme/mq_app_theme.dart';
+// import 'package:flutter/material.dart';
+// import 'package:mq_app_theme/mq_app_theme.dart';
 
-@immutable
-final class AppLocalDataSourceMock implements AppLocalDataSource {
-  const AppLocalDataSourceMock();
+// @immutable
+// final class AppLocalDataSourceMock implements AppLocalDataSource {
+//   const AppLocalDataSourceMock();
 
-  @override
-  String appVersion() => '1.1.1';
+//   @override
+//   String appVersion() => '1.1.1';
 
-  @override
-  CustomTheme initialTheme() => const CustomTheme(Brightness.light, AppColors.red);
+//   @override
+//   CustomTheme initialTheme() => const CustomTheme(Brightness.light, AppColors.red);
 
-  @override
-  Future<void> saveThemeMode({required bool isDark}) => Future.value();
+//   @override
+//   Future<void> saveThemeMode({required bool isDark}) => Future.value();
 
-  @override
-  Future<void> saveThemePrimaryColor(int index) => Future.value();
-}
+//   @override
+//   Future<void> saveThemePrimaryColor(int index) => Future.value();
+// }
