+++ packages/mq_app_theme/lib/src/source/mock/read_theme_local_mock_data_source.dart
@@ -1,17 +1,17 @@
-import 'package:flutter/material.dart';
-import 'package:mq_app_theme/mq_app_theme.dart';
+// import 'package:flutter/material.dart';
+// import 'package:mq_app_theme/mq_app_theme.dart';
 
-@immutable
-final class LocalThemeDataSourceMock implements LocalThemeDataSource {
-  const LocalThemeDataSourceMock();
+// @immutable
+// final class LocalThemeDataSourceMock implements LocalThemeDataSource {
+//   const LocalThemeDataSourceMock();
 
-  @override
-  ReadThemeData getInitialThemeState() {
-    return ReadThemeData.initial;
-  }
+//   @override
+//   ReadThemeData getInitialThemeState() {
+//     return ReadThemeData.initial;
+//   }
 
-  @override
-  Future<void> saveThemeState(ReadThemeData themeState) {
-    return Future.value();
-  }
-}
+//   @override
+//   Future<void> saveThemeState(ReadThemeData themeState) {
+//     return Future.value();
+//   }
+// }
