+++ packages/mq_app_theme/lib/src/source/read_theme_local_data_source.dart
@@ -1,7 +1,7 @@
-import 'package:mq_app_theme/mq_app_theme.dart';
+// import 'package:mq_app_theme/mq_app_theme.dart';
 
-abstract class LocalThemeDataSource {
-  ReadThemeData getInitialThemeState();
+// abstract class LocalThemeDataSource {
+//   ReadThemeData getInitialThemeState();
 
-  Future<void> saveThemeState(ReadThemeData themeState);
-}
+//   Future<void> saveThemeState(ReadThemeData themeState);
+// }
