+++ packages/mq_app_ui/lib/src/mq_app_ui.dart
@@ -1,3 +1,4 @@
+export 'repo/repo.dart';
 export 'colors/app_colors.dart';
 export 'extensions/date_time_extension.dart';
 export 'spacing/app_spacing.dart';
@@ -6,7 +7,6 @@ export 'typography/app_font_weight.dart';
 export 'typography/app_text_styles.dart';
 export 'generated/assets.gen.dart';
 export 'generated/fonts.gen.dart';
-export 'notifier/theme_notifier.dart';
 export 'shape/custom_track_shape.dart';
 export 'decoration/gradient_decorated_box.dart';
 export 'enum/gender_enum.dart';
