+++ packages/mq_app_ui/lib/src/repo/repossitories/app_repository.dart
@@ -1,0 +1,11 @@
+import 'package:mq_app_ui/mq_app_ui.dart';
+
+abstract class AppRepository {
+  String appVersion();
+
+  AppTheme getInitialThemeState();
+
+  Future<void> saveThemePrimaryColor({required bool isOrange});
+
+  Future<void> saveThemeMode({required bool isDark});
+}
