+++ packages/mq_app_ui/lib/src/repo/sources/app_local_data_source.dart
@@ -1,0 +1,11 @@
+import 'package:mq_app_ui/mq_app_ui.dart';
+
+abstract class AppLocalDataSource {
+  String appVersion();
+
+  AppTheme initialTheme();
+
+  Future<void> saveThemePrimaryColor({required bool isOrange});
+
+  Future<void> saveThemeMode({required bool isDark});
+}
