+++ packages/mq_app_ui/lib/src/repo/sources/mock/read_theme_local_mock_data_source.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:mq_app_ui/mq_app_ui.dart';
+
+@immutable
+final class LocalThemeDataSourceMock implements LocalThemeDataSource {
+  const LocalThemeDataSourceMock();
+
+  @override
+  ReadThemeData getInitialThemeState() {
+    return ReadThemeData.initial;
+  }
+
+  @override
+  Future<void> saveThemeState(ReadThemeData themeState) {
+    return Future.value();
+  }
+}
