+++ packages/mq_app_ui/lib/src/theme/app_theme.dart
@@ -117,4 +117,7 @@ abstract class AppTheme {
   ChipThemeData get chipTheme {
     return const ChipThemeData();
   }
+
+  bool get isOrange => colorScheme.primary == AppColors.tomato;
+  bool get isDark => colorScheme.brightness == Brightness.dark;
 }
